% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normconf.R
\name{get_talk}
\alias{get_talk}
\title{Find out what someone's talk will be about according to GPT-2}
\usage{
get_talk(title)
}
\arguments{
\item{title}{A talk title}
}
\description{
Find out what someone's talk will be about according to GPT-2
}
