% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normconf.R
\name{schedule}
\alias{schedule}
\title{Normconf Schedule Goodness}
\usage{
schedule()
}
\description{
Retrieves and prints the schedule from the schedule endpoint.
You may want to have your R console set nice and wide for this one.
}
